+++ dbt/adapters/clickhouse/nativeclient.py
@@ -68,6 +68,8 @@ class ChNativeClient(ChClientWrapper):
             client_name=f'dbt-adapters/{dbt_adapters_version} dbt-clickhouse/{dbt_clickhouse_version} clickhouse-driver/{driver_version}',
             secure=credentials.secure,
             verify=credentials.verify,
+            certfile=credentials.client_cert,
+            keyfile=credentials.client_cert_key,
             connect_timeout=credentials.connect_timeout,
             send_receive_timeout=credentials.send_receive_timeout,
             sync_request_timeout=credentials.sync_request_timeout,